приложение/backend/requirements.txt
flask==2.3.2
psycopg2-binary==2.9.6
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
faker==18.11.2